requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
defusedxml==0.6.0
Django==2.2
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==19.9.0
idna==2.10
importlib-metadata==3.4.0
install==1.3.4
kombu==4.6.11
meld3==2.0.1
oauthlib==3.1.0
Pillow==8.1.0
prompt-toolkit==3.0.16
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
supervisor==4.0.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
